// frontend/src/app/clients/melderegister/model/countryDTO.ts
/**
 * Zentrales Melderegister
 * Zentrales Melderegister
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CountryDTO { 
    coordinatesLatitude?: number;
    coordinatesLongitude?: number;
    id?: number;
    name?: string;
    stateId?: number;
}